requirements.txt
flask~=3.0.3
python-dotenv
ruff
pip-autoremove
pymongo
gunicorn